Moving DataRepository to repository package
