+++ app/src/main/java/com/c23ps419/petanikita/di/Injection.kt
@@ -1,7 +1,7 @@
 package com.c23ps419.petanikita.di
 
 import android.app.Application
-import com.c23ps419.petanikita.data.DataRepository
+import com.c23ps419.petanikita.data.repository.DataRepository
 import com.c23ps419.petanikita.data.local.datastore.UserPreferences
 import com.c23ps419.petanikita.data.remote.network.ApiConfig
 import com.c23ps419.petanikita.data.repository.UserDatabaseRepository
