+++ app/src/main/java/com/c23ps419/petanikita/ui/login/LoginViewModel.kt
@@ -1,7 +1,7 @@
 package com.c23ps419.petanikita.ui.login
 
 import androidx.lifecycle.ViewModel
-import com.c23ps419.petanikita.data.DataRepository
+import com.c23ps419.petanikita.data.repository.DataRepository
 
 class LoginViewModel(private val dataRepository: DataRepository): ViewModel() {
     fun userLogin(email: String, password: String) = dataRepository.postLogin(email, password)
