+++ app/src/main/java/com/c23ps419/petanikita/ui/register/RegisterViewModel.kt
@@ -1,7 +1,7 @@
 package com.c23ps419.petanikita.ui.register
 
 import androidx.lifecycle.ViewModel
-import com.c23ps419.petanikita.data.DataRepository
+import com.c23ps419.petanikita.data.repository.DataRepository
 
 class RegisterViewModel(private val dataRepository: DataRepository): ViewModel() {
     fun userRegister(name: String, email: String, phone: String, password: String) = dataRepository.postRegister(name, email, phone, password)
